backend/requirements.txt
asgiref==3.8.1 ; python_version >= "3.13"
django-cors-headers==4.7.0 ; python_version >= "3.13"
django-filter==25.1 ; python_version >= "3.13"
django==5.2.1 ; python_version >= "3.13"
djangorestframework-simplejwt==5.5.0 ; python_version >= "3.13"
djangorestframework==3.16.0 ; python_version >= "3.13"
gunicorn==23.0.0 ; python_version >= "3.13"
packaging==25.0 ; python_version >= "3.13"
psycopg-binary==3.2.9 ; python_version >= "3.13" and implementation_name != "pypy"
psycopg-pool==3.2.6 ; python_version >= "3.13"
psycopg==3.2.9 ; python_version >= "3.13"
pyjwt==2.9.0 ; python_version >= "3.13"
python-dotenv==1.1.0 ; python_version >= "3.13"
sqlparse==0.5.3 ; python_version >= "3.13"
typing-extensions==4.13.2 ; python_version >= "3.13"
tzdata==2025.2 ; python_version >= "3.13" and sys_platform == "win32"
